+++ CHANGELOG.md
@@ -17,6 +17,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [0.6.2] - 2025-01-12
+
+### Added
+
+- Refresh button for Subscription usage
+- Prompt history in Chat Input
+
+### Changed
+
+
+### Fixed
+
+- Location for log and run files in Linux
+- Subscription page updates when navigating away and back
+
+
 ## [0.6.1] - 2025-01-12
 
 ### Added